# apps/corpus/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin corpus
  version: "1.0.0"
  description: Per-DID corpus indexing and BM25 search service — ingests threads from local workspace files and GitHub repositories, and serves evidence-based search over them.
servers:
  - url: http://localhost:8003
    description: internal (not publicly routed; consumed only by the kernel MCP proxy)

tags:
  - name: corpus

components:
  securitySchemes:
    ImajinClaim:
      type: apiKey
      in: header
      name: Authorization
      description: >-
        `Authorization: Imajin-Claim <base64url-encoded claim>.<ed25519 signature>`.
        A short-lived (60s, 5-minute hard ceiling) CorpusAccessClaim minted by
        the kernel (apps/kernel/src/lib/kernel/corpus-access-claim.ts) after its
        own corpus:read/corpus:write scope gate passes, using the kernel's
        existing node DID key (AUTH_PRIVATE_KEY) — no new keypair. The claim is
        scoped to exactly the :did in the request path and carries alg, aud,
        scope, issuedAt, expiresAt, and a nonce. Corpus verifies the signature
        against an env-pinned public key (CORPUS_KERNEL_PUBLIC_KEY — corpus
        never holds or derives the kernel's private key), plus alg, aud,
        expiry, nonce replay, and that the claim's DID matches :did, before
        serving any /corpus/:did/* route (apps/corpus/src/middleware/access-claim.ts).
        Rejects with 401 on missing/invalid/expired/wrong-alg/wrong-aud claims,
        403 on a DID mismatch.
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }

paths:
  /health:
    get:
      tags: [corpus]
      operationId: health
      summary: Liveness check for the corpus service.
      description: Public and unauthenticated by design — used for process/container health probes only.
      security: []
      responses:
        "200":
          description: Service is up
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  service: { type: string }

  /spec:
    get:
      tags: [corpus]
      operationId: getSpec
      summary: Serve this OpenAPI spec (YAML) so discovery tooling can enumerate corpus's routes.
      description: Public and unauthenticated by design, matching every sibling service's /spec (or /{service}/api/spec) route.
      security: []
      responses:
        "200":
          description: OpenAPI YAML spec
          content:
            text/yaml:
              schema: { type: string }

  /corpus/{did}/status:
    get:
      tags: [corpus]
      operationId: corpusStatus
      summary: Report indexed sources, thread counts, and freshness for a DID's corpus.
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:read to this :did; minted by the kernel's corpus_status MCP tool / dashboard client.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Corpus status
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources: { type: array, items: { type: object } }
                  threadCount: { type: integer }
                  pendingEmbeddings: { type: integer, description: "Embedding chunks awaiting (or re-awaiting, after a failed PGX attempt) an embed pass (#1599, #1601)." }

  /corpus/{did}/search:
    post:
      tags: [corpus]
      operationId: corpusSearch
      summary: Search a DID's corpus for threads matching a query, returning scored evidence excerpts.
      description: >-
        Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:read to this :did; minted by the kernel MCP corpus_search tool.
        Set `ref` to pin results to a previously-ingested git sha for a `local:workspace` source (#1921): the response
        is then filtered to that ref's snapshot manifest and includes a `provenance` block (sha, doc ids, content hashes).
        Set `mode` (#1599, #1601) to force a retrieval strategy; defaults to `hybrid` when a PGX embedder is configured
        for this service, `bm25` otherwise. A `ref`-pinned query always runs `bm25`-only regardless of `mode`, since the
        vector index has no "as of this ref" notion.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                sourceType: { type: string }
                source: { type: string, description: "Required when ref is set." }
                state: { type: string }
                type: { type: string }
                labels: { type: array, items: { type: string } }
                author: { type: string }
                limit: { type: integer }
                budget: { type: integer }
                ref: { type: string, description: "Pin the query to a previously-ingested git sha for source (#1921). Requires source." }
                mode: { type: string, enum: [bm25, semantic, hybrid], description: "Force a retrieval strategy (#1599, #1601). See the operation description for defaults." }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results: { type: array, items: { type: object } }
                  totalHits: { type: integer }
                  freshness: { type: array, items: { type: object } }
                  tokensUsed: { type: integer }
                  provenance: { type: object }
                  mode: { type: string, enum: [bm25, semantic, hybrid], description: "Retrieval strategy actually used to produce results." }
                  degraded:
                    type: array
                    items: { type: string, enum: [semantic, rerank] }
                    description: >-
                      Present only when a requested capability silently fell back instead of erroring —
                      e.g. ["semantic"] when PGX embedding failed and results are BM25-only, or ["rerank"]
                      when the PGX reranker was unavailable but semantic retrieval still ran.
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Unknown ref — no snapshot has been indexed for this `(source, ref)` pair. Trigger ingest at this ref, never a silent fallback to HEAD.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/ingest:
    post:
      tags: [corpus]
      operationId: corpusIngest
      summary: Ingest raw ThreadDocument[] directly, or crawl+ingest a `{ source }` (local:workspace or github:owner/repo).
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:write to this :did; minted by the kernel MCP corpus_load tool.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
                  required: [source]
                  properties:
                    source: { type: string }
      responses:
        "200":
          description: Ingest result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/sources:
    post:
      tags: [corpus]
      operationId: corpusRegisterSource
      summary: Register a source (`{ source, type }`) and perform its initial full crawl — the GitHub adapter's entry point (#1729).
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:write to this :did. Not yet proxied by a kernel MCP tool — a caller must mint its own claim directly today.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string, description: "e.g. \"local:workspace\" or \"github:owner/repo\"" }
                type: { type: string, description: "Optional; must match the source's own prefix, e.g. \"github\"." }
      responses:
        "200":
          description: Registration + initial crawl result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/sync:
    post:
      tags: [corpus]
      operationId: corpusSync
      summary: Incrementally sync a previously-registered `{ source, cursor }` (local:workspace or github:owner/repo).
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:write to this :did; minted by the kernel MCP corpus_sync tool.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string }
                cursor: { type: string, nullable: true }
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
                  cursor: { type: string, nullable: true }
                  hasMore: { type: boolean }
        "501":
          description: Source is unset or not a recognized (workspace/github) source
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/crawl:
    post:
      tags: [corpus]
      operationId: corpusCrawl
      summary: Full crawl+ingest of a `{ source }` (local:workspace or github:owner/repo), without registering it.
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:write to this :did. Not yet proxied by a kernel MCP tool — a caller must mint its own claim directly today.
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [source]
              properties:
                source: { type: string }
      responses:
        "200":
          description: Crawl result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ingested: { type: integer }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /corpus/{did}/source:
    delete:
      tags: [corpus]
      operationId: corpusDeleteSource
      summary: Delete every indexed thread for a `{ source }` from a DID's corpus.
      description: Requires a CorpusAccessClaim (see ImajinClaim) scoped corpus:write to this :did; minted by the kernel's corpus-client.ts (dashboard delete flow).
      security:
        - ImajinClaim: []
      parameters:
        - name: did
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source: { type: string }
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted: { type: integer }